Move client code to client module
